requirements.txt
Brotli==1.0.7
click==7.1.2
cycler==0.10.0
dash==1.13.4
dash-bootstrap-components==0.10.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
IMDbPY==6.8
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
pandas==1.0.5
plotly==4.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
scipy==1.5.0
seaborn==0.10.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
